+++ src/main/java/ru/yandex/practicum/filmorate/storage/feed/FeedLine.java
@@ -25,19 +25,18 @@ import java.time.Instant;
 @Slf4j
 @RequiredArgsConstructor
 public class FeedLine {
-
     private final JdbcTemplate jdbcTemplate;
     private final FeedDbStorage feedDbStorage;
 
-    private void insertData(int idUser, int idEntity, String event, String operation) {
+    private void insertData(Long idUser, Long idEntity, String event, String operation) {
         final String sqlQuery = "INSERT INTO feed (USER_ID, ENTITY_ID, EVENT_TYPE, OPERATION, TIMELONG) " +
                 "VALUES(?, ?, ?, ?, ?) ";
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update((connection) -> {
             PreparedStatement statement = connection.prepareStatement(sqlQuery, new String[]{"EVENT_ID"});
-            statement.setInt(1, idUser);
-            statement.setInt(2, idEntity);
+            statement.setLong(1, idUser);
+            statement.setLong(2, idEntity);
             statement.setString(3, event);
             statement.setString(4, operation);
             statement.setLong(5, Instant.now().toEpochMilli());
@@ -46,25 +45,25 @@ public class FeedLine {
     }
 
     @AfterReturning("execution(* ru.yandex.practicum.filmorate.service.FriendService.addFriend(..)) && args(id, friendId))")
-    public void afterReturningCallAtAddFriend(int id, int friendId) {
+    public void afterReturningCallAtAddFriend(Long id, Long friendId) {
         log.info("//////Лента новостей////// Пользователь id_1={} добавил в друзья пользователя id_2={}", id, friendId);
         insertData(id, friendId, EventType.FRIEND.toString(), Operation.ADD.toString());
     }
 
     @AfterReturning("execution(* ru.yandex.practicum.filmorate.service.FriendService.deleteFriend(..)) && args(id, friendId))")
-    public void afterReturningCallAtDeleteFriend(int id, int friendId) {
+    public void afterReturningCallAtDeleteFriend(Long id, Long friendId) {
         log.info("//////Лента новостей////// Пользователь id_1={} удалил пользователя id_2={} из друзей", id, friendId);
         insertData(id, friendId, EventType.FRIEND.toString(), Operation.REMOVE.toString());
     }
 
     @AfterReturning("execution(* ru.yandex.practicum.filmorate.service.LikeService.addLike(..)) && args(filmId, userId))")
-    public void afterReturningCallAtAddLikeFilm(int filmId, int userId) {
+    public void afterReturningCallAtAddLikeFilm(Long filmId, Long userId) {
         log.info("//////Лента новостей////// Пользователь id_1={} добавил лайк к фильму id_2={}", userId, filmId);
         insertData(userId, filmId, EventType.LIKE.toString(), Operation.ADD.toString());
     }
 
     @AfterReturning("execution(* ru.yandex.practicum.filmorate.service.LikeService.deleteLike(..)) && args(filmId, userId))")
-    public void afterReturningCallAtDeleteLikeFilm(int filmId, int userId) {
+    public void afterReturningCallAtDeleteLikeFilm(Long filmId, Long userId) {
         log.info("//////Лента новостей////// Пользователь id_1={} удалил лайк от фильма id_2={}", userId, filmId);
         insertData(userId, filmId, EventType.LIKE.toString(), Operation.REMOVE.toString());
     }
@@ -76,7 +75,7 @@ public class FeedLine {
     }
 
     @AfterReturning("execution(* ru.yandex.practicum.filmorate.service.ReviewService.delete(..)) && args(idReview))")
-    public void afterReturningCallAtDeleteReview(int idReview) {
+    public void afterReturningCallAtDeleteReview(Long idReview) {
         log.info("//////Лента новостей////// Ревью id_1={} пользователя id_2={} удалено", idReview, feedDbStorage.getIdAuthor(idReview));
         insertData(feedDbStorage.getIdAuthor(idReview), idReview, EventType.REVIEW.toString(), Operation.REMOVE.toString());
     }
